+++ apps/graphslam-engine/CMakeLists.txt
@@ -18,8 +18,6 @@ project(${PROJECT_NAME})
 
 set(SRC_FILES
     graphslam-engine_app.cpp
-    TSlidingWindow.h TSlidingWindow.cpp
-    CEdgeCounter.h CEdgeCounter.cpp
     )
 
 # Define the executable target:
