+++ libs/graphslam/CMakeLists.txt
@@ -1,12 +1,18 @@
 #---------------------------------------------
 # Macro declared in "DeclareMRPTLib.cmake":
 #---------------------------------------------
-define_mrpt_lib_header_only(
+define_mrpt_lib(
 	# Lib name
 	graphslam
 	# Dependencies
-	mrpt-graphs
 	mrpt-base
+	mrpt-graphs
+	mrpt-gui
+	mrpt-hwdrivers
+	mrpt-maps
+	mrpt-obs
+	mrpt-opengl
+	mrpt-slam
 	)
 
 IF(BUILD_mrpt-graphslam)
