+++ libs/graphslam/include/mrpt/graphslam/CEmptyERD.h
