+++ libs/graphslam/include/mrpt/graphslam/CEmptyNRD.h
