+++ libs/graphslam/include/mrpt/graphslam/CFixedIntervalsNRD.h
