+++ libs/graphslam/include/mrpt/graphslam/CGraphSlamOptimizer.h
@@ -19,7 +19,7 @@
 #include <mrpt/utils/TParameters.h>
 #include <mrpt/utils/CTimeLogger.h>
 
-#include "CWindowManager.h"
+#include <mrpt/graphslam/CWindowManager.h>
 #include "CRegistrationDeciderOrOptimizer.h"
 
 namespace mrpt { namespace graphslam { namespace optimizers {
