+++ libs/graphslam/include/mrpt/graphslam/CICPCriteriaERD_impl.h
