+++ libs/graphslam/include/mrpt/graphslam/CICPCriteriaNRD.h
