+++ libs/graphslam/include/mrpt/graphslam/CICPCriteriaNRD_impl.h
