+++ libs/graphslam/include/mrpt/graphslam/CLoopCloserERD.h
