+++ libs/graphslam/include/mrpt/graphslam/CLoopCloserERD_impl.h
