+++ libs/graphslam/include/mrpt/graphslam/CNodeRegistrationDecider.h
@@ -15,12 +15,12 @@
 #include <mrpt/obs/CSensoryFrame.h>
 #include <mrpt/obs/CObservation.h>
 #include <mrpt/gui/CDisplayWindow3D.h>
+#include <mrpt/graphslam/CWindowManager.h>
 #include <mrpt/graphs/CNetworkOfPoses.h>
 #include <mrpt/synch/CCriticalSection.h>
 #include <mrpt/utils/TParameters.h>
 #include <mrpt/utils/CTimeLogger.h>
 
-#include "CWindowManager.h"
 #include "CRegistrationDeciderOrOptimizer.h"
