+++ libs/graphslam/include/mrpt/graphslam/CRangeScanRegistrationDecider.h
