+++ libs/graphslam/include/mrpt/graphslam/CRangeScanRegistrationDecider_impl.h
