+++ libs/graphslam/include/mrpt/graphslam/CRegistrationDeciderOrOptimizer.h
