+++ libs/graphslam/include/mrpt/graphslam/TSlidingWindow.h
