+++ libs/graphslam/src/CEdgeCounter.cpp
@@ -1,4 +1,13 @@
-#include "CEdgeCounter.h"
+/* +---------------------------------------------------------------------------+
+   |                     Mobile Robot Programming Toolkit (MRPT)               |
+   |                          http://www.mrpt.org/                             |
+   |                                                                           |
+   | Copyright (c) 2005-2016, Individual contributors, see AUTHORS file        |
+   | See: http://www.mrpt.org/Authors - All rights reserved.                   |
+   | Released under BSD License. See details in http://www.mrpt.org/License    |
+   +---------------------------------------------------------------------------+ */
+
+#include <mrpt/graphslam/CEdgeCounter.h>
 
 using namespace mrpt::graphslam;
